Separate TimerLoader to its own component
